correct rounding

+++ core/ilwisobjects/domain/numericrange.h
@@ -84,7 +84,7 @@ public:
     {
         double value = v.toDouble();
         if ( _resolution != 0.0)
-             value = (double)(value / _resolution) * _resolution;
+             value = (double)((qint64)(value / _resolution)) * _resolution;
         if ( !contains(value, inclusive))
             return _undefined;
